Abstracted out the DataFile elements from the CSVFile class - so we can reuse these for the XMLFile and JSONFile class
